updated itemcontroller to use sessions

+++ BackEnd/app/src/main/java/org/campusmarket/app/controllers/ItemController.java
@@ -29,12 +29,9 @@ import org.springframework.web.server.ResponseStatusException;
 
 
 @RestController
-//@RequestMapping("/items")
-
-
-
-
-public class ItemController {
+@RequestMapping("/items")
+public class ItemController 
+{
 	@Autowired
     private ItemsRepository items;
 	
@@ -46,25 +43,24 @@ public class ItemController {
 
 	
 	Log log = LogFactory.getLog(ItemController.class);
-
 	
-	
-	@RequestMapping("/items/all")
-	public List<Item> getAll(){
-		try {
-			
-			
-	    return items.findAll();
+	@RequestMapping("/all")
+	public List<Item> getAll()
+	{
+		try 
+		{
+	    	return items.findAll();
 		}
-		   catch (Exception e){
+		catch (Exception e)
+		{
             log.error(e.getMessage());
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No items were found.");
         }
     }
 
 	 
-	@PostMapping("/users/{username}/items/new")
-	public void newItem(@RequestBody Item item, @PathVariable("username") String username, @RequestParam(name = "sessid", required = true) String sessid)
+	@PostMapping("/new")
+	public void newItem(@RequestBody Item item, @RequestParam(name = "sessid", required = true) String sessid)
 	{
 		if (sessid.isEmpty())
         {
@@ -84,105 +80,114 @@ public class ItemController {
         log.info(" success: a new item was created with a reference number(keep for your record): " + item.getRefnum());
 
 		}
-		   catch (Exception e){
-	            log.error(e.getMessage());
-	            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "There's no such user with this username so sorry we won't be able to add your item :(",e);
-	        }
-	    }
-	
-	
-    @PutMapping("/users/{username}/items/{refnum}")
-	public void updateItem(@RequestBody Item item, @PathVariable (value = "username") String username, 
-			 @PathVariable (value = "refnum") int refnum) {
-    
-    	try {
-    		
-    		if(users.existsByUserName(username)==0) {
-    			 
-    			log.error("There's no such user with this username in our system so sorry we can't update this item");
-    			return;
-
-    		}
-    		  		
-    	
-    	Item oldItem=items.findByRefnum(refnum);
-    	User u= users.findByUsername(username);
-
-		if(u.getAdmin() || oldItem.getUser().getUsername().equals(username))
+		catch (Exception e)
 		{
-			oldItem.setName(item.getName());
-			oldItem.setPrice(item.getPrice());
-			oldItem.setCategory(item.getCategory());
-			oldItem.setCondition(item.getCondition());
-			items.save(oldItem);
-			
-			log.info(" success: the item with a reference number of " + refnum +" was updated");
-			return;
-    	}
-    	
-    	log.error("Unathorized attempt: Could not update the item(you don't own the item) with refnum: " + refnum);
-    	
-    	} catch (Exception e) {
-    		 log.error(e.getMessage());
-    		throw new ResponseStatusException(HttpStatus.BAD_REQUEST, " Failure: Could not update the item with refnum: " + refnum);
-    		
-        }
-
-    }
-    
-    
-    
-    @RequestMapping(value = "/users/{username}/items/{refnum}/delete", method = RequestMethod.DELETE)
-    public void deleteItem(@PathVariable (value = "username") String username, 
-			 @PathVariable (value = "refnum") int refnum)
-    {
-    	
-		try 
-		{	
-        	
-			if(users.existsByUserName(username)==0)
-			{
-   			 
-    			log.error("Unathorized attempt: There's no such user with this username in our system so sorry we can't delete this item");
-    			return;
-
-    		}
-        	
-        	Item oldItem=items.findByRefnum(refnum);
-        	User u= users.findByUsername(username);
-
-        	if(u.getAdmin() || oldItem.getUser().getUsername().equals(username)){
-        		
-        		items.deleteById(refnum);
-                log.info(" success: the item with a reference number of " + refnum +" was deleted");
-                return;
-        	}
-        	log.error("Unathorized attempt: Could not delete the item(you don't own the item) with refnum: " + refnum);
-        	
-        }
-        catch (Exception e)
+	        log.error(e.getMessage());
+	        throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "There's no such user with this username so sorry we won't be able to add your item :(",e);
+	    }
+	}
+	
+	
+    @PutMapping("/update/{refnum}")
+	public void updateItem(@RequestBody Item item, 
+							@PathVariable (value = "refnum") int refnum,
+							@RequestParam(name = "sessid", required = true) String sessid) 
+	{
+		if (sessid.isEmpty())
         {
-            log.error("item with refnum"+ refnum + " not found" );
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "item with refnum " + refnum + " not found.", e);
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Request Invalid: Empty value for required parameter 'sessid'.");
         }
+
+        Session active = sessions.findBySessId(sessid);
+        
+        if (active == null) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Could not find an active session with id: " + sessid);
+
+		User user = users.findById(sessions.findUserBySession(sessid));
+
+		if (user == null) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Could not find user.");
+
+		Item oldItem = items.findByRefnum(refnum);
+
+		if (oldItem == null) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Could not find item with refnum: " + refnum);
+
+		if (active.getAdmin() || item.getUser().getUsername().compareTo(user.getUsername()) == 0)
+		{
+			try 
+			{
+				oldItem.setName(item.getName());
+				oldItem.setPrice(item.getPrice());
+				oldItem.setCategory(item.getCategory());
+				oldItem.setCondition(item.getCondition());
+				items.save(oldItem);
+					
+				log.info(" success: the item with a reference number of " + refnum +" was updated");
+			} 
+			catch (Exception e)
+			{
+				log.error(e.getMessage());
+				throw new ResponseStatusException(HttpStatus.BAD_REQUEST, " Failure: Could not update the item with refnum: " + refnum);
+			}
+		}
+		else throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Access denied.");
+
     }
     
-    @GetMapping("/users/{username}/items/sellers")
-    public ArrayList<Item> findItemBySeller(@PathVariable("username") String seller) {
-    	
-    	try {
-    	
-    	return items.findBySeller(seller);
-    	
-    	} catch(Exception e) {
-    		log.error(e.getMessage());
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No item with this seller was found in the cymarket");
+    
+    
+    @RequestMapping(value = "/delete/{refnum}", method = RequestMethod.DELETE)
+	public void deleteItem( @PathVariable(value = "refnum") int refnum,
+							@RequestParam(name = "sessid", required = true) String sessid)
+    {
+    	if (sessid.isEmpty())
+        {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Request Invalid: Empty value for required parameter 'sessid'.");
+        }
 
+        Session active = sessions.findBySessId(sessid);
+        
+        if (active == null) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Could not find an active session with id: " + sessid);
+
+		User user = users.findById(sessions.findUserBySession(sessid));
+
+		if (user == null) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Could not find user.");
+
+		Item oldItem = items.findByRefnum(refnum);
+
+		if (oldItem == null) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Could not find item with refnum: " + refnum);
+		
+		if (active.getAdmin() || oldItem.getUser().getUsername().compareTo(user.getUsername()) == 0)
+		{
+			try 
+			{
+				items.deleteById(refnum);
+				log.info(" success: the item with a reference number of " + refnum +" was deleted");
+				
+			}
+			catch (Exception e)
+			{
+				log.error("item with refnum"+ refnum + " not found" );
+				throw new ResponseStatusException(HttpStatus.NOT_FOUND, "item with refnum " + refnum + " not found.", e);
+			}
+		}
+		else throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Access denied.");
+    }
+    
+    @GetMapping("/sellers/{username}")
+	public ArrayList<Item> findItemBySeller(@PathVariable("username") String seller)
+	{	
+		try 
+		{
+    		return items.findBySeller(seller);
+		}
+		catch(Exception e)
+		{
+    		log.error(e.getMessage());
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "This user has no items for sale.");
     	}
     }
      
            
-    @GetMapping("/items/name/{name}")
+    @GetMapping("/name/{name}")
     public Collection<Item> findItemByName(@PathVariable("name") String name) {
     	
     	try {
@@ -195,7 +200,7 @@ public class ItemController {
     	}
     }
      
-    @GetMapping("/items/category/{category}")
+    @GetMapping("/category/{category}")
     public Collection<Item> findItemByCategory(@PathVariable("category") String category) {
     	try {
     	return items.findByCategory(category);
@@ -209,7 +214,7 @@ public class ItemController {
     
     
     
-    @GetMapping("/items/cond/{cond}")
+    @GetMapping("/cond/{cond}")
     public Collection<Item> findItemByCondition(@PathVariable("cond") String cond) {
     	try {
     	return items.findByCond(cond);
@@ -223,7 +228,7 @@ public class ItemController {
  
     
      
-    @GetMapping("/items/name/{name}/cond/{cond}")
+    @GetMapping("/name/{name}/cond/{cond}")
     public Collection<Item>findByCondAndName(@PathVariable("name") String name ,@PathVariable("cond") String cond){
     	try {
 		return items.findByCondAndName(name, cond);
@@ -234,7 +239,7 @@ public class ItemController {
     	}
     }
     
-    @GetMapping("/items/category/{category}/cond/{cond}/price/{price}")
+    @GetMapping("/category/{category}/cond/{cond}/price/{price}")
     public Collection<Item>findByCondPriceCategory(@PathVariable("cond") String cond ,@PathVariable("category") String category, @PathVariable("price") double price){
     	try {
 		return items.findByCondAndCategoryAndPrice(cond, category, price);
@@ -244,17 +249,31 @@ public class ItemController {
     	}
     }
     
-    @RequestMapping(value = "/items/delete/all", method = RequestMethod.DELETE)
-    public void deleteAll()
+    @RequestMapping(value = "/delete/all", method = RequestMethod.DELETE)
+    public void deleteAll(@RequestParam(name = "sessid", required = true) String sessid)
     {
-    	try {
-    	
-         items.deleteAll(); 
-         log.info("User Table Cleared: all users removed.");
+		if (sessid.isEmpty())
+        {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Request Invalid: Empty value for required parameter 'sessid'.");
+        }
 
-    } catch (Exception e) {
-    	log.error("No items in database to remove.");
-        throw new ResponseStatusException(HttpStatus.NO_CONTENT, "No items in database to remove.");
-    }
+        Session active = sessions.findBySessId(sessid);
+        
+        if (active == null) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Could not find an active session with id: " + sessid);
+
+		if (active.getAdmin())
+		{
+			try
+			{
+				items.deleteAll(); 
+				log.info("User Table Cleared: all users removed.");
+			} 
+			catch (Exception e) 
+			{
+				log.error("No items in database to remove.");
+				throw new ResponseStatusException(HttpStatus.NO_CONTENT, "No items in database to remove.");
+			}
+		}
+		else throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Access denied.");
    }
 }
